// apps/api/src/lib/mesh/__generated/sdk.ts
export type AgentSummaryFragment = { agentURI: string, createdAt: string, createdAtTransaction: string, id: string, feedbackCount: string, owner: { address: string } | null, profile:
    | { name: string | null, description: string | null, image: string | null }
    | { name: string | null, description: string | null, image: string | null }
   | null, metadata: Array<{ key: string, value: string }> };